libipt, block: provide instruction class of last instruction if available
In most cases the block decoder knows the instruction class of the instruction
that ended the block.  Provide it to save our callers the effort of decoding the
instruction in some cases.

If it doesn't know the instruction class, set it to ptic_error and leave it to
our callers to decide whether they need to decode the last instruction.

Change-Id: I4174b628377635ccce0d5d8a8ebe71853592782c
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>