libipt: classify VMX instruction from VM's perspective
We classify VMLAUNCH/VMRESUME as far call and VMCALL as far return.  This is
from a VMM's perspective that is calling into guests.

This is the opposite of how we handle OS calls where we treat SYSENTER as far
call and SYSEXIT as far return.

Change the VMX classification to align the two, i.e. classify VMLAUNCH/VMRESUME
as far return and VMCALL as far call.

Change-Id: I1a914d6c26a5807f820038390cfa3c0847b84071
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>