libipt: classify software interrupts as far branches
Software interrupts (INT, INT1, INT3, INTO) are used for software breakpoints
and for system calls.  The former are typically patched in and removed again
before trace decode.  The latter remain in the code.  They are currently
classified as 'other'.

Classify software interrupts as far calls similar to SYSCALL and SYSENTER.

Change-Id: I8aedaa8bded0215e04c3b797445ffc866700d6c9
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>