ptxed, block: determine error ip
When diagnosing block decode or control-flow reconstruction errors, we report
"block.ip + block.ninsn" as IP, where the latter is omitted if it is zero.

Change this to use the current IP while printing for diagnosing control-flow
reconstruction errors and block.end_ip or the next IP for diagnosing block
decode errors depending on the erro type.

This improves error reporting when --quiet is specified and further frees us
from having to update block.ip and block.ninsn during printing.

Change-Id: I906bdd18411e092019f19de40bc7bd51162fa225
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>