libipt: timing events
Add a new event type ptev_tick that provides an update to the estimated Time
Stamp Count.  The event will be used to provide timing updates that are not
bound to other events.

The event time reflects the time at the last consumed packet and thus the time
at the last instruction (in the last block).  The intended use-case is to
correlate the instruction stream with timestamped events such as trace sideband
information.

Change-Id: I835d13063fdf178c1e697d40013e733749bc2f58
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>