libipt, block: tick events
Add a new block decoder configuration flag enable_tick_events to generate
ptev_tick events on conditional or indirect branch instructions.

The precision of timing information depends on the trace configuration.  If
cycle accurate tracing is enabled, the time should correspond to the last
instruction in the block that indicated the event.

Change-Id: If3fa121cdf3de044f6ebe641c27a6de7888db24b
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>