libipt, block: always forward enabled events
Tracing enabled events are sometimes indicated by setting the enabled flag in
struct pt_block and sometimes by forwarding the event to ther user.

Always forward the event and process it in pt_blk_event().  Once we did this for
all events, it will simplify both the decoder and the API.

Change-Id: Ibf5a5e75c7d98816c1af52f24acca035898ed7a0
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>