libipt, block: make pt_blk_proceed_(no_)event() handle trailing events
In pt_blk_collect() we proceed by one block, then process trailing events.  The
latter may indicate events to be forwarded to the user.  This requires us to
determine whether an event shall be forwarded to the user by examining the event
and the decoder state in pt_blk_process_trailing_events().

For some events, e.g. ptev_ptwrite, this requires us to indicate the event in
pt_blk_proceed_event() by setting a flag in the decoder state.  If we wanted to
forward more events, e.g. ptev_disabled, we could follow this route.  It is a
bit clumsy, though.  It would be nicer if we could already indicate a user event
in pt_blk_proceed_event().

This is a first step in that direction.

With the call to pt_blk_process_trailing_events() moved into pt_blk_proceed(),
pt_blk_collect() has become an otherwise empty wrapper around pt_blk_proceed().
Remove it.

Change-Id: I4f0828dbceb8822c0122ac36c808774560a66eab
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>