ptxed: change order of [interrupt] and [aborted] notifications
On a transaction abort due to an interrupt both the interrupt and aborted flags
will be set and ptxed will print:

    [aborted]
    [interrupt]

On the packet level, this corresponds to MODE.TSX + FUP + TIP(.PGD), which is
translated into a ptev_tsx event followed by a ptev_async_branch event.  This is
the other way round.

If we're going to forward the above events to ptxed instead of indicating them
via pt_insn or pt_block flags, we will get them in the oppositer order:

    [interrupt]
    [aborted]

Change the order in which ptxed prints those two notifications and all tests
that use them to prepare for such a change.

Change-Id: I2cb546ecdc27c56ee08fb5a7c33144b9675da370
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>