libipt: forward tsx events to the user while tracing is disabled
We silently ignore ptev_tsx events while tracing is disabled.  We use them to
update the decoder state but don't tell the user about them arguing that there
can be a real lot of them and that they are not particularly helpful.

That's not really for us to decide.  Forward those events to the user.

Change-Id: Ifaaac5549eabe0378bc096a406376fabe33dcd1f
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>