libipt, block: avoid trailing event detour for power events
Power events are already forwarded to the user.  If we are able to reach the
event location in pt_blk_proceed_event(), there's no need to go through
pt_blk_process_trailing_events() in order to signal the event to the user.

Directly call pt_blk_status() to indicate the event.

Change-Id: Iabe8396b7d9dcdf2395dc852bc4ef358c69ab018
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>