libipt, block: avoid trailing event detour for ptwrite events
A ptwrite event is forwarded to the user.  To avoid consuming multiple ptwrite
events on the same PTWRITE instruction, we track whether we reached the PTWRITE
instruction in pt_blk_proceed_event() in order to indicate it in
pt_blk_process_trailing_events().

Since we are now able to indicate events directly in pt_blk_proceed_event(), we
no longer need to keep track of whether we already forwarded the event for the
current instruction.  It suffices to handle ptwrite events encountered while
tracing is disabled in pt_blk_process_trailing_events().

We lose the ability to diagnose erraneous pt_blk_event() calls without a
previous indication in the pt_blk_next() status return for ptwrite.  This is
already the case for other events, e.g. ptev_overflow, ptev_stop.  Once we
forward all events to the user and rely on calls to pt_blk_event() to process
them, we won't get very far if this switching between pt_blk_next() and
pt_blk_event() is not working reliably.

Also discard a check that ptwrite events can't be status update events.  This
check is alredy missing for other events, e.g. ptev_exstop, ptev_pwre.

Change-Id: I5b8c36b32f5f64a1d72d9b52b13bce36fc5809f3
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>