libipt, insn: user events proceed in after event processing
Event processing after decode allows forwarding the event to the user.  This
will skip peek event processing but it will still proceed to the next
instruction.  The only event that makes use of this is ptev_ptwrite.

Move the proceed step into event processing in case an event is forwarded to the
user.  This allows forwaring other events we can't proceed, e.g. ptev_disabled.

Change-Id: I5549c79f1b769b3c40034e59774c2a77b53d5856
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>