libipt, insn: always forward tsx events
TSX events are sometimes indicated by setting the committed and aborted flags in
struct pt_insn and sometimes by forwarding the event to the user.

Always forward the event to the user and process it in pt_insn_event().  We keep
indicating speculative execution by setting the speculative flag in struct
pt_insn.

Change-Id: Icec084169d1e7b44fffaae37cb8a57d3b3a10f81
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>