libipt, insn: remove process_events_before()
We now forward all events to the user.  Events that used to be processed before
decoding the instruction in process_events_before() are now indicated in
process_events_peek() and processed in pt_insn_event().

This removes error diagnostics that might catch cases where the user forgot to
call pt_insn_event() and instead called pt_insn_next().  Such a bug is quite
fundamental, though, and we wouldn't get very far - probably not even past the
initial events in PSB+ and to the first instruction.

Change-Id: I90c20c9f7fd2db32166b99db6009671f0aa09ef2
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>