libipt, insn: remove enabled check before pt_insn_proceed()
We check whether tracing is enabled at the beginning of pt_insn_next().  If it
got disabled due to a synchronous disable event binding to the current
instruction, we'd indicate the event to the user and skip pt_insn_proceed().

If we reach pt_insn_proceed(), then, tracing must still be enabled.  Remove the
check.

Change-Id: Ic16590b41f6c90853685a96a10b8f78547265b0b
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>