libipt, insn: remove last_disable_ip
We store the IP at which we expect tracing to resume in last_disable_ip so we
can then compare it to the actual IP at which tracing was re-enabled.  We don't
really need a separate variable for that.

Use the decoder's ip field, which is otherwise unused while tracing is disabled.

Change-Id: I77932a257c55e0d3b3c6da0600131fe237a9f109
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>