libipt, insn: copy the decoded instruction earlier
We provide the decoded instruction to the user at the end of pt_insn_next().
This was necessary as long as we indicated events via flags in struct pt_insn.
Now that we forward all events to the user, the instruction object won't change
during event processing.

This allows us to move the insn_to_user() call right after pt_insn_decode(),
which in turn allows further simplifications.

Change-Id: Ic5b5422770809a20940f2732f8263f223252ebac
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>