ptdump: support perf_event sideband
Add new options

  --pevent <file>[:<from>[-<to>]]
  --pevent:primary <file>[:<from>[-<to>]]
  --pevent:secondary <file>[:<from>[-<to>]]

to dump the Linux perf_event sideband contained in <file>.  An optional offset
or range can be given as <from> and <to> file offsets.  The sideband records
are interleaved with the trace packets based on TSC.

Add new options

  --pevent:sample-type
  --pevent:time-zero
  --pevent:time-shift
  --pevent:time-mult
  --pevent:tsc-offset
  --pevent:kernel-start
  --pevent:sysroot
  --pevent:kcore
  --pevent:vdso-x64
  --pevent:vdso-x32
  --pevent:vdso-ia32

to configure sideband iterpretation.

This feature requires the SIDEBAND and the PEVENT cmake options to be ON.

Change-Id: I35afd8d775a15f70cc3d93aa37c9a2d8f53a3b09
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>