ptxed: support perf_event sideband
Add new options

  --pevent:primary <file>[:<from>[-<to>]]
  --pevent:secondary <file>[:<from>[-<to>]]

to add a Linux perf_event sideband decoder working on the sideband stream in
<file>.  An optional file offset or range can be given as <from> and <to> file
offsets.  Use

  --pevent:primary <file>[:<from>[-<to>]]

for sideband files that are directly connected to the trace (e.g. that provide
the sideband for the traced cpu) and

  --pevent:secondary <file>[:<from>[-<to>]]

for sideband files that are indirectly connected to the trace (e.g. that provide
sideband for other cpus).

Add new options

  --pevent:sample-type
  --pevent:time-zero
  --pevent:time-shift
  --pevent:time-mult
  --pevent:tsc-offset
  --pevent:kernel-start
  --pevent:sysroot
  --pevent:kcore
  --pevent:vdso-x64
  --pevent:vdso-x32
  --pevent:vdso-ia32

to configure the perf_event sideband decoder.

Use --pevent: options to change the configuration for the next --pevent:primary
or --pevent:secondary option.

This feature requires the SIDEBAND and the PEVENT cmake options to be ON.
The --pevent:kcore option also requires the FEATURE_ELF cmake option to be ON.

Change-Id: Ia164463b11421ac2218a2395e0ab88ef3498f92a
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>