libipt, query: last event time
When querying for the current time using pt_qry_time(), we return the time at
the decoder's current position.  Since the decoder is reading ahead in order to
indicate upcoming events, the returned time can be significantly in the future.

We therefore recommend to read the time before querying for an indirect or
conditional branch.  This works fine for indirect branches.  Since conditional
branches are served out of the TNT cache, though, the time will still be in the
future starting from the second branch in a TNT packet.

To address that, we store the current time before reading ahead and return that
time in pt_qry_time().  The provided time will now be in the past instead of in
the future and the recommendation changes to query the current time after
querying for an indirect or conditional branch.

Change-Id: Iedda3612d1c50f56a385bdaafb41be6735ec7cfa
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>