pttc: remove local re-declarations of variables
In p_gen_expfile() and parse_ip() we re-declare a few variables in a nested
block scope that are already defined on function scope.  Those variables are not
alive across that block scope so we can use the function scope version, instead.

Also rename a local variable errstr in p_get_expfile() into errtext to avoid
shadowing the global errstr array.

This avoids a warning with VS2015.

Change-Id: Id015b832dea4af8e375e75b7e8774476d54c92b7
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>