ptunit: change error type when string-comparing NULL pointers
When using a NULL pointer in ptunit string comparison, ptunit fails with a
pointer error rather than a string error.

Change this to fail with a string error and report the NULL string as "(null)".

The trigger for this change was to address a compiler warning about shadowed
variables from VS2015.  Beyond that, I believe that the changed error type
better reflects the actual problem.

Change-Id: Iafe92e096d8884f2032cd1686b2d37d180ae2fe2
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>