cmake: snprintf
We used to define snprintf to _snprintf_c for Windows builds in the top-level
cmakefile.  That function is meanwile defined so we only need that workaround
for older version.

Remove it from the cmakefile and instead define it in the files that use
snprintf based on _MSC_VER.

This fixes #16.

Change-Id: Id3775ffc1a1c29a257c4062d3cfbe153198449de
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>