libipt, ptxed: add cbr events
Add a new event type ptev_cbr for frequency changes.

This requires changing the way we search for FUP packets when decoding OVF
packets.

We remove the OVF-cutoff-FUP unit test to allow treating the incomplete OVF+FUP
as a stand-alone OVF at the end of the trace.  We could detect the truncated FUP
and preserve the old behaviour but we'd lose the information that the trace
ended with an overflow.

Change-Id: Icf62d846a335d8f2418de33f41cd4febfb5d3874
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>