pttc: redo ipc parsing
For parsing IP compression, we first store a long int into an pt_ip_compression
enum and later check the value.  This does not cover very large values that are
transformed into valid IPC values when converted to int.  Change it to detect
errors during conversion from the integer input to the pt_ip_compression enum.

This fixes a clang -Wbad-function-cast diagnostic.

Change-Id: I010d69b510d66f9030963f4258aa0a48c78cd150
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>