libipt: indicate events at end_on_call IP
When a block ends at a call instruction due to the end_on_call configuration
option, we forgot to indicate events at that IP.  On the next call to
pt_blk_next(), we would return an empty block and indicate the event.

Check whether we reached an event IP on end_on_call to avoid the empty block.

Change-Id: I506e51e0457826c56049a7097cfd76809614d7bc
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>