libipt: check file offset for otherwise identical image section overlaps
When mapping different sections of the same file into the same virtual address
range, pt_image_add_file() would incorrectly assume that the two mappings are
identical and ignore the addition/replacement.

There is special support for ignoring the repeated addition of the same file
section.  This checks the virtual address range, the image section identifier,
and the filename.  It forgot to check the file offset.  Add it.

Change-Id: Ic632d30d1cfb46369c1780d33c76f9e415b4c9bb
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>