libipt: remove special case for identical image section adds
When adding an image section, there's a special check whether we're adding an
existing section again.  This avoids unmapping and re-mapping that section.

With the image section cache, that will be handled by use-counting.  There is no
need for such an explicit check in pt_image_add_*().  Remove it.

This now requires a check for copy-to-self in pt_image_copy() since moving
sections around would otherwise interfere with the iteration over sections for
adding them one-by-one.

Change-Id: Ic35c9b9aac671fa8e1dc0a8fe6d11461ee285bec
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>