libipt, query: return offset when searching for FUP after OVF
pt_qry_find_ovf_fup() returns 1 to indicate that it found a FUP after an OVF.
Extend this to return the offset from the current position (after the OVF) to
(after the) FUP.  We will need this in future patches.

We can not return the offset to the FUP packet since that might be zero, which
would indicate that no FUP was found.

Change-Id: I414287278a6e9ffc2df086fb74db534393126c97
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>