libipt: ignore partial status packets at end of trace
When an incomplete status packet, e.g. MTC or CYC, is encountered at the end of
the trace, the upcoming end is not indicated correctly by setting the pts_eos
status bit.  We rely instead on the -pte_eos error to be repeated on the next
query.

This may cause decoders to decode too far, for example when the trace is split
for multi-threaded decode.  It may also cause an incorrect -pte_no_enable error
instead of the expected -pte_eos when the trace ends with tracing disabled.

Change-Id: If1b521329b58f121d412b4519dcbceb67e4d90eb
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>