iscache: maintain pointer from section to iscache
For image sections that are added to an image section cache, store a pointer to
the iscache that created the section in the section object.

The pointer is initialized when the iscache creates the section and cleared when
the iscache removes the section.  As long as the iscache pointer is non-NULL,
the iscache holds a reference to the section.

The internal API for this is a pair of functions to attach and detach an image
section cache to a section.  It would allow moving a section from one image
section cache to another if we ever wanted to.  There can be at most one image
section cache attached to an image section at any time.

Change-Id: I7ff07057580701a4f4af767e72ea1209ae988280
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>