libipt, iscache: add mapped section cache
Add a cache of recently used sections to an image section cache.  The image
section cache will keep the sections in the cache mapped even if the image that
mapped the section unmaps it again.

This prevents repeated unmapping and remapping in some parallel decode
scenarios and when reading from the image section cache.

A new API function pt_iscache_set_limit() can be used to set the memory limit of
the cache in bytes.  A value of zero disables caching.  A non-zero value
specifies the amount of memory that can be used by the cache.  By default, the
cache is unlimited.

Change-Id: Ib71da0c6b96dd9c6553d3ae75bc0123baf845aa5
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>