libipt: request block cache in block decoder
When an image section is mapped, a block cache is implicitly allocated for that
section.  For image sections that are only used by instruction flow decoders,
the block cache will not be used.

Allocate the block cache on request from the block decoder.

We also add a separate flow to the iscache's mapped sections cache to avoid
having to consider section size changes on the common flow.

Change-Id: I38445c043eb2ec5035ab05180e9392fdc67f7379
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>