libipt, image: synthesize on-map notifications
Without a cache in the image, sections would get mapped when they are used and
unmapped again when they are not used.  The image keeps sections mapped even
though they are not currently used.

When cycling through the sections cached in an image, there will only be a
single pt_section_map() call when the section is added to the cache.  To an
attached image section cache it looks like those sections are not used any
longer.

Synthesize calls to pt_section_on_map() whenever a section is moved to the front
of the image's cache to keep the caches in sync.

Change-Id: Id124f93124f713f24ecbafae52c38ea433fc6e3f
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>