libipt, image, block: rework section splitting
When adding a new section to an image, any existing sections that overlap with
the new section are split and new sections are added for the remaining bytes.

This is done by creating a new section for the same file using different file
offsets resulting in smaller sections and hence less memory to map.

With the new mapped sections cache in the image section cache it makes more
sense to re-use the original section as it can be shared between decoders.

This will also preserve the block cache.  Block cache entries made after the
section was split may cause additional stops for decoders sharing the block
cache and using the original section.

This introduces a potential issue when a section is split in the middle of a
sequence of linear instructions that ends with an unconditional direct near
branch back into the sub-section that also contained the start of the sequence.

When the middle of this sequence is overwritten and changed in the process, the
block cache may still contain an entry that stays entirely in one sub-section
and we may not be able to detect the section split.  This would result in a
decode error.

Change-Id: Ibbd3134e4a42a2ecb1cf0e62ce8c22241f7e2d7e
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>