libipt, image: extract reading from a mapped section
Extract a function to read from a mapped section from pt_image.c into
pt_mapped_section.h.

Add a check to ensure that the virtual address we're reading from lies in the
mapped section and truncate the read to remain in that section.

Redundant checks will be addressed in subsequent commits.

Change-Id: I9e381658883af3d64e937fb49fc2118034025003
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>