libipt, block: insert an artificial stop on direct near backwards branches
The previous patch left a corner case where splitting a section in the middle of
a linear sequence of instructions that jumped back to the originating section
might leave a stale block cache entry.

We avoid this by inserting an artificial decode stop for direct near backwards
branches.  This allows us to detect section switches in such cases.

Leaving this as a separate commit to allow measuring the overhead.

Change-Id: Ib42b638c79ddbcd5a079ff14c09a63c1c74f07c1
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>