libipt, block: end-on-jump
Add a new block decoder flag

    end_on_jump

that ends a block after every direct near jump.

This is similar to the existing end_on_call block decoder flag and allows faster
detection of tail calls since only the last instruction in a block needs to be
considered as tail call candidate.

Change-Id: Ib81dbd95e6dcc5470fcae29541c19d2b2427bcd5
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>