libipt: fix synchronous disable regression
Commits

    1767865 libipt: skl014 (partial)
    6fff5a9 libipt: skl014

added a workaround for skl014 where a TIP.PGD may not get an IP payload if
FilterEn were cleared on a direct unconditional branch.

Without a filter configuration the workaround disables tracing on the first
direct unconditional branch after a synchronous disable with suppressed IP.

This conservatively assumes that the erratum may have hit.  On the other hand,
not having a filter configuration may also mean that no address filters were
used.  In this case, we may disable tracing too early.

Change the assumption to assume that no address filters were used unless they
are specified in the decoder configuration.

Change-Id: I3c8786623fb6dd09bc0c1caf716a3965a12c3eee
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>