ptxed: add option to only load trace from a coredump
The options

  --core:<task> <file>[:<from>[-<to>]]
  --elf <file>[:<base>]

can be used to load the traced memory image from an ELF file.

In addition to loading the traced memory image, the former will also load the
trace for the given task.

Add another option

  --core-pt:<task> <file>[:<from>[-<to>]]

to load the trace for the given task but not the traced memory image.

This can be used to split

  --core:<task> <file>[:<from>[-<to>]]

into

  --core-pt:<task> <file>[:<from>[-<to>]] --elf <file>

which is quite useful for debugging.

Change-Id: I7a66f35ce00474bb3e80bfc46e6162164a7c32c6
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>