libipt: preserve error when retrying instruction decode fails to read memory
When we fail to decode an instruction at the end of a section where the memory
read returned less than the requested number of bytes, we retry by reading the
remaining bytes and decode again.

If there is no more memory available that read fails with -pte_nomap, which we
return as result of the instruction decode.

Change this to preserve the original -pte_bad-insn that triggered the retry.

Change-Id: I5c62ed8b967aae3c2adf2fd27ceff31e0e906d3f
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>