libipt, ptxed, ptdump, pttc: diagnose unknown cpus
When configuring errata for a cpu using pt_cpu_errata() there is no indication
whether the cpu is known and errata have been configured.  The may cause decode
fails later on if the cpu wasn't known.

Fail with -pte_bad_cpu if the cpu is not known.

If the --cpu option is used for ptdump, ptxed, or pttc and the cpu has not been
set to none, pttc will fail and ptdump and ptxed will indicate the error in
their regular output.

Change-Id: Ibe114c498e55e1054a28124e190a3e1692a55e84
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>