libipt, iscache: avoid add/map deadlock
There is a potential deadlock when we're adding a section that matches an
existing section that is currently in use.

  - when adding a section, we lock the iscache and then attach to the
    to-be-added section, which requires taking the section's attach lock

  - when mapping a section, we take the section's attach lock to notify
    an attached iscache to update its LRU, which requires locking the iscache

We avoid the deadlock by temporarily releasing the iscache lock in order to
attach to the candidate section.  This opens a window where the iscache may be
modified and we'd want to share a different section.  We may therefore need to
repeat this step.

Change-Id: I6019bec8becbe05265d81f6de5c8bc97708e4145
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>