libipt: sync at partially consumed PSB
In case of packet decode errors, the next PSB may have been partially
attributed to the previous packet - which then led to the packet decode
error.

When synchronizing onto the trace stream in forward direction, start
sizeof(PSB) - 1 bytes before the current decoder position.

When synchronizing in backward direction, we are already considering the
trace around the current decoder position.

Change-Id: I753f12503d4c79736fa57084e44fdc90bfafdd36
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>