pttc: check for string length underflows
When renaming the sideband file, we add the configuration to the filename
between the basename and the file extension.  The original filename should
already contain the extension.

We subtract the extension length from the filename length in order to strip
the extension from the filename.  This may underflow if the filename did not
contain the extension.  Check that case.  We will end up with a corrupted
filename in that case.

Change-Id: I4f39829ff5f5f7f68d7bc642b250c530f9ae0038
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>