cmake: fix add_cflag_if_available()
In add_cflag_if_available(), we compute a guard variable from the
to-be-checked option by appending _supported to it.  Cmake passes the guard
variable as macro to the compiler.  And GCC 8 complains that macros must be
identifiers.  This causes the check to always fail.

Fix it by passing the name of the guard variable to add_cflag_if_available().
It is up to the caller to ensure unique names.

Change-Id: I78ee4779b58b9dfe74bfe93f5d309909f7c6a442
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>