libipt, ptdump, ptxed: optionally preserve timing calibration on OVF
At an OVF packet, we do not know what packets were lost.  If we lost any
timing packets, calibration, and hence timing, may be significantly off for
the current and the next MTC periods.

To avoid backwards jumps in timing, we clear our calibration state and
restart calibrating.  Calibration will be up-to-date again after CBR or
after two MTC.

Until then, we ignore CYC packets.  We still update timing on TSC and MTC.
This results in fewer timing updates but time should always move forward.

Add a new decoder configuration flag keep_tcal_on_ovf to preserve the
current calibration result as we reset calibration.  This results in the
usual frequency of timing updates on every CYC.

We should be behind the cycles we lost in this MTC period but there's still
a risk of overcounting, which would result in a backwards jump at this or
the next MTC, when we correct calibration errors.

Change-Id: If0d4c58694c5fb1ed8c9be55388b61a256da914b
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>