libipt: allow adding bits to the tnt cache
The TNT cache currently represents a single TNT packet.  It is updated from
a TNT packet; updating requires that the cache is empty.

Change this to allow TNT bits to be added to a non-empty cache.

Also rename the update function to pt_tnt_cache_add() to reflect that and
replace the TNT packet parameter with two parameters for the bits to add and
their size.

Change-Id: I8fffc936108b9fa8ce2b24be29ce1ea40a0e6d46
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>