doc: using xed from github
In doc/howto_build.md we point the reader to the XED files when using XED
from a PIN distribution.

Add a pointer to the include files and to XED's README when using XED from
github.  This also clarifies that XED needs to be built before it can be
used.

Change-Id: I7d84a21948ee136e5a0ea55a7e8c5df4f0e990fa
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>