libipt, blk: no tnt cache
With the new event decoder, the TNT bits are stored in a TNT event and we
can use them from there.  There is no need to further cache TNTs.

The tricky part is to place tick events.  We currently overwrite the TNT or
TIP event with the tick event.  But we can only do that once the event is no
longer used.

The timestamp on a TNT event corresponds to the first bit in the TNT so we
want to place it after the conditional branch corresponding to that first
TNT bit.  The event may contain further TNT bits so we cannot overwrite it.

We instead indicate the tick event on the TNT event itself and synthesize
the tick event in pt_blk_event().

Change-Id: I803b7c828de6d1d4eb5fd3c688dfea4aa36eafe8
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>