ptunit: restrict file names to FILENAME_MAX characters
In case of errors, ptunit prints the basename of the executable test file
and the line number of the failing check.

Restrict the length of the complete filename to FILENAME_MAX characters.
For filenames exceeding that length, the first FILENAME_MAX characters of
the complete filename are printed.

This should not really occur in practice.

Change-Id: I188544d84d615b008e49a752b5fdfe5e96bd3470
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>