pttc: avoid reading the yasm output text twice
When processing the yasm output text string, we first determine the length
of the string using strlen() and then iterate over the text to split it into
lines.

We already know the size, however, since the caller allocated the buffer
containing the string.

As a side-effect, we now treat embedded NUL bytes as empty lines.  They used
to terminate processing.

Change-Id: I86187fd8c7b98aec64d23b6303cb050d555fcbda
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>