ptunit, section: try to remove file while holding a handle to it
Some systems allow removing a temporary file while holding a handle to it.
We prefer doing that since it should be the safest option we have.

Some systems do not allow removing open files.  Remove the temporary file we
created after closing it if the first attempt to remove it fails.  Given
that those are temporary files with temporary file names, clashes should be
very rare.  It appears better than leaking the file.

Change-Id: Ia0d013598567cb4089071e4678cc3e8d0e3e00b2
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>