libipt, ptxed: rename ptic_error to ptic_unknown
In some cases, we intentionally do not provide an instruction classification
for performance reasons.  Or we did not get that far due to errors.

In both cases, the instruction classification itself is not an error; the
instruction class is simply not known.  If we fail to length-decode or
classify an instruction, we fail with a negative pt_error_code.

Rename the instruction class

    ptic_error

into

    ptic_unknown

Fixes #73.

Change-Id: I8fdb5971876b17661c0996c85f13c1eb656eb267
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>