ci: example workflow for Coverity scan
Unfortunately, Coverity does not allow to do in-place pull request
analysis with an immediate verification. Build results have to be
uploaded to the Coverity server for further analysis and report
generation, and projects using free license may get throttled if they
submit it too often. It seems to be a common pattern to do scans on
master/main branch daily and check reports sent to a specified mail.

Signed-off-by: Mihails Strasuns <mihails.strasuns@intel.com>