ptdump, ptxed, pttc: drop --cpu auto
The sample tools offer an option

    -- cpu auto

to determine the family/model/stepping of the processor they are running on,
i.e. on which trace is decoded.

This may lead to confusion as decode requires the family/model/stepping of
the processor on which trace was recorded.

When recording with Linux perf, for example, this information is stored in
the perf.data file, from which script/perf-get-opts.bash extracts it.

Drop the option to avoid confusion.

Change-Id: I2c572a7e540a7577faeeff80408d8fb5d04ef950
Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>