all: bump minor version
We used to bump the version number when tagging a release and use the commit
message to provide a summary of changes.

This leaves master with the same major.minor version as the last stable
branch and users without a means for checking the version on master.

Reverse the process and bump the version on master after tagging a release
to have master already report the new version.  The version may still change
on master if we introduce non-backwards-compatible changes later one.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>