libipt: use atomic operations to access section->bcache
The block cache is lazily allocated under the section lock and only deleted
under the section lock when the map count drops to zero.  It is accessed
without taking the lock, however, relying on atomic load behavior on IA and
most other architectures, which triggers a thread sanitizer warning.

Use atomic operations for accessing the block cache.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>