libipt, evq: merge event queues
The event queue consists of a collection of queues, one per binding packet
type.  The interface takes a binding packet type and uses the corresponding
queue for the request.  This allows only a single binding packet per event.

Extend the interface to take a bitmask of binding packet types.  Any entry
binds to a certain packet type if the corresponding bit is set in its
binding bitmask.  This allows an event to bind to more than one packet type.

Merge all those per-binding packet type queues into a single queue and add
the vector of potential binding packets to each entry.  This keeps events in
chronological order.

Signed-off-by: Markus Metzger <markus.t.metzger@intel.com>